Begin to implement indirect drawing

+++ src/Voxels/world/Mesher.hpp
@@ -10,8 +10,8 @@ bool inBounds(int x, int y, int z, int size, int height);
 
 int dirToIndex(int i, int j, int k);
 
-uint64_t createVertex(int x, int y, int z, int colour, int normal, int ao);
+uint32_t createVertex(int x, int y, int z, int colour, int normal, int ao);
 
-void meshChunk(Chunk *chunk, int worldSize, std::vector<uint64_t> &data);
+void meshChunk(Chunk *chunk, int worldSize, std::vector<uint32_t> &data);
 
 bool boundsCheck(int x, int y, int z, int i, int j, int k, int worldSize, std::vector<int> &voxels);
